% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{convert_alex_database_med_1}
\alias{convert_alex_database_med_1}
\title{a function used to code antiotics, Biologics, Enteral.Steroids, en and ppi as yes, no, past use.}
\usage{
convert_alex_database_med_1(medRaw, medName, period)
}
\arguments{
\item{medRaw:}{a data frame that contains abx, Biologics, en, Enteral.Steroids, ppi information, contains ehb_id instead of nautilusid}

\item{medName:}{the name of the medication variable}

\item{period:}{used to define the current usage of that medication and the unit is days. for example, if period = 10, then the current usage will be defined as within 10 days (10 included) of the sample collect date}
}
\value{
medConverted: a data frame that contains yes or no, one column for one medication
}
\description{
a function used to code antiotics, Biologics, Enteral.Steroids, en and ppi as yes, no, past use.
}
