% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{join_fromalex_sample}
\alias{join_fromalex_sample}
\title{return all rows and all columns from both data frames. Where there are not matching values, returns NA for the one missing.}
\usage{
join_fromalex_sample(sample_df1, sample_df2, ID = "sample_subject_name")
}
\arguments{
\item{sample_df1}{the 1st data frame}

\item{sample_df2}{the 2nd data frame to be joind}

\item{ID}{names of the columns used to merge the two data frames, default set to sample_subject_name}
}
\description{
return all rows and all columns from both data frames. Where there are not matching values, returns NA for the one missing.
}
\examples{
# load sample_visit_event
alex_sample_visit_event <- read_alex_database(fileName = "/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/sample_visit_event.csv", uselessColumns = c("X"))
alex_tobacco_grant_albumin <- read_alex_database(file = "//Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/tobacco_grant_albumin.csv", uselessColumns = c("X", "index","rank_id")) \%>\% unique
alex_sample <- join_fromalex_sample(alex_sample_visit_event, alex_tobacco_grant_albumin)
}
