% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{join_fromalex_subject}
\alias{join_fromalex_subject}
\title{Join two dataframe that contains subject level information together by using certain columns (default set to ebh_id, redcap_event_name)
return all rows and all columns from both data frames. Where there are not matching values, returns NA for the one missing.}
\usage{
join_fromalex_subject(subject_df1, subject_df2, ID = c("ehb_id",
  "redcap_event_name"))
}
\arguments{
\item{subject_df1}{the 1st data frame}

\item{subject_df2}{the 2nd data frame to be joind}

\item{ID}{names of the columns used to merge the two data frames, default set to ehb_id, redcap_event_name}
}
\description{
Join two dataframe that contains subject level information together by using certain columns (default set to ebh_id, redcap_event_name)
return all rows and all columns from both data frames. Where there are not matching values, returns NA for the one missing.
}
\examples{
alex_demographics_information <- read_alex_database(fileName ="/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/demographics_information.csv")
alex_antiobiotics_timing <- read_alex_database(fileName = "/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/antiobiotics_timing.csv")
alex_subject <- join_fromalex_subject(alex_demographics_information, alex_antiobiotics_timing)

}
