% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{process_alex_database_med_log}
\alias{process_alex_database_med_log}
\title{use convert_alex_database_med_2 to convert and combine all the medRaw data}
\usage{
process_alex_database_med_log(medRaw)
}
\arguments{
\item{medRaw:}{a data frame that contains abx, Biologics, en, Enteral.Steroids, ppi information, contains ehb_id instead of nautilusid}
}
\value{
medData a data frame in wide format sample in rows and medication in columns.
}
\description{
use convert_alex_database_med_2 to convert and combine all the medRaw data
}
