% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{read_alex_database}
\alias{read_alex_database}
\title{Read data from Alex's database}
\usage{
read_alex_database(fileName, uselessColumns = c("X", "index"))
}
\arguments{
\item{fileName}{name of the file to read in}

\item{uselessColumns}{useless columns to be removed, default set to X and index}
}
\value{
alexData a data frame
}
\description{
Read data from Alex's database
}
\examples{
alexData = read_alex_database(fileName = "/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/external_records_protocols.csv")


}
