% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alexdata.R
\name{read_alex_database_med_log}
\alias{read_alex_database_med_log}
\title{function used to clear up the medication log data, keep only the medication in the medList, remove the unused records.}
\usage{
read_alex_database_med_log(dataFrame, medList)
}
\arguments{
\item{dataFrame}{the raw medication log data frame contains the data from baseline_antibiotics.csv and medication_log.csv and a column of collect date time}
}
\value{
medRaw return a data frame that contains only the medications in the medList
}
\description{
function used to clear up the medication log data, keep only the medication in the medList, remove the unused records.
}
\examples{
medList <- as.list(read.xlsx("/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromsarah/IBD Medication List_2.17.2017.xls", sheetIndex = 1,stringsAsFactors = F))
medList <- sapply(medList, function(x) x = x[!is.na(x)])
names(medList)[which(names(medList) == "Other")] = "Other.med"
names(medList)[which(names(medList) == "X5.ASA")] = "5.ASA"
alex_medication_log <- read_alex_database(fileName = "/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/medication_log.csv")
alex_medication_log \%<>\% mutate(start_date = as.Date(start_date)) \%>\%
mutate(stop_date = as.Date(stop_date)) \%>\%
mutate(duration = stop_date - start_date)
a <- read_alex_database(fileName = "/Users/yue/Dropbox/Joy/tabacco_grant/data/data_fromalex/database/database10052017/sample_redcap.csv") \%>\%
rename(collect_date_time = Collected_Time)
alex_medication_log \%<>\% left_join(a, by = c("ehb_id")) \%>\%
rename(SampleType = Sample.Type)
remove(a)
medRaw = read_alex_database_med_log(dataFrame = alex_medication_log, medList = medList)
}
